начал стили для профиля, импортировал sass node-sass
