--- frontend/src/content/components/ProfileBox/styles.scss
@@ -1,3 +1,0 @@
-.profile-box {
-  border: 1px solid blue;
-}
