+++ frontend/src/view/components/Content/ContentFeed/ContentFeed.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Search from "../Search/Search";
+import Search from "../../Search/Search";
 import ContentGrid from "../ContentGrid/ContentGrid";
 
 const ContentFeed = ({addContentButton, data}) => {
