+++ frontend/src/view/components/Content/ContentGrid/ContentGrid.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
-import Post from "../Post/Post";
+import Post from "../../Post/Post";
+import styles from './styles.module.scss.css'
 
 const ContentGrid = ({data}) => {
     console.log(data)
     return (
-        <div>
+        <div className={styles.flexContainer}>
             {data.map(postInfo =>
                 <Post key={postInfo.id} data={postInfo}/>
             )}
