+++ frontend/src/view/components/Content/ContentGrid/styles.module.scss.css
@@ -1,0 +1,10 @@
+.box {
+    display: flex;
+    flex-direction: row;
+}
+
+.flexContainer {
+    /*генерирует flex-контейнер уровня блока*/
+    display: -webkit-flex;
+    display: flex;
+}
