+++ frontend/src/view/components/ProfileBox/profile.module.scss
@@ -1,0 +1,88 @@
+.profileBox {
+  border: 1px solid blue;
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  padding: 0px;
+}
+
+.avatarBox {
+  border: 1px solid green;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding: 20px 0px;
+
+  width: 150px;
+  height: 190px;
+
+
+  /* Inside auto layout */
+
+  flex: none;
+  order: 0;
+  align-self: stretch;
+  flex-grow: 0;
+}
+
+.mainInfoBox{
+  border: 1px solid red;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding: 20px 0px 20px 20px;
+  gap: 20px;
+
+  width: 1006px;
+  height: 190px;
+
+
+  /* Inside auto layout */
+
+  flex: none;
+  order: 1;
+  align-self: stretch;
+  flex-grow: 1;
+}
+
+.row1 {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding: 0px;
+
+}
+
+.row2 {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  padding: 0px;
+  gap: 20px;
+
+}
+
+.row3 {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  padding: 8px 0px 0px;
+  gap: 10px;
+
+}
+
+.nickname {
+  display: flex;
+  flex: 1 0 100%;
+  flex-direction: column;
+  align-items: flex-start;
+  padding: 0px;
+}
+
+.statistics {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  padding: 0px;
+  gap: 20px;
+}
