+++ frontend/src/view/components/Search/Search.jsx
@@ -3,7 +3,8 @@ import React from 'react';
 const Search = ({addContentButton}) => {
     return (
         <div>
-
+            <input placeholder={'Поиск...'}/>
+            {addContentButton && <button>Добавить работу</button>}
         </div>
     );
 };
