+++ frontend/src/view/pages/private/chamber/Collaborations.jsx
