+++ frontend/src/view/pages/private/chamber/People.jsx
